requirements.txt
streamlit
pytesseract
pdf2image
pillow
ollama
faiss-cpu
langchain
